A new tool to query nodes and their hardware properties at chassis level and register the node information in the Ironic database.
This tool targets chassis management of micro-servers or similar enlosures
where the chassis has an active interface to query the details of large
number of baremetal nodes. Scaling the number of nodes is also managed with
a view that querying individual nodes to inspect their properties will consume
significant time.

This tool will reside in the ironic/tools folder.